AtomicCms Scaffold Generator
Why:

In order to easily manage content on presentable pages, a generator is
needed to do the initial legwork of creating and editing content.

This change addresses the need by:

Creating a scaffold generator to create a migration, model, controller,
route, view, and admin resource for creating, displaying, and editing
content.