Generalize form selectors
Why?

Currently the javascript only supports editing one type of model, a
"page"

How?

Updated the angular code to look for a form field by id: cms_type
If this exists, we set the selectors to use that cms_type as a prefix,
otherwise, we stay with the default, "page"

Side Effects?

I left "page" in there as to support the current functionality, so
hopefully there shouldn't be anything unanticipated. But, there are no
gurantees here with javascript :)