Add the ability to nest components in the CMS
Why:

We need to be able to include some component inside of other bigger
components and have them render a dynamic number of options.

This change addresses the need by:

Including some methods from the ArrayComponent class in the larger
Editor class so they are available to all inheriting classes.
Update the atomic_cms.js to be able to add subcomponents to parent
ones.